+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All significant modifications to this project will be recorded in this file. The
 
 ## [Unreleased]
 
+## [1.2.0] - 2025-08-18
+
+### Added
+
+- Reading + Meaning mode.
+- 'Audio Initialization' stage.
+
 ## [1.1.0] - 2025-01-12
 
 ### Added